% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_helpers.R
\name{cohens_h_to_prop}
\alias{cohens_h_to_prop}
\title{Calculate Two Proportions from Cohen's h}
\usage{
cohens_h_to_prop(h, p2)
}
\arguments{
\item{h}{Cohen's h effect size}

\item{p2}{Proportion in group 2 (0 to 1)}
}
\value{
Proportion in group 1
}
\description{
Inverse transformation: given p2 and Cohen's h, calculate p1
}
\keyword{internal}
