% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_components.R
\name{create_advanced_settings}
\alias{create_advanced_settings}
\title{Create Advanced Settings Components}
\usage{
create_advanced_settings()
}
\description{
Create Advanced Settings Components
}
\keyword{internal}
