% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test_registry.R
\name{create_correlation_spec}
\alias{create_correlation_spec}
\title{Correlation Test Specification}
\usage{
create_correlation_spec()
}
\description{
Correlation Test Specification
}
\keyword{internal}
