% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_components.R
\name{create_effect_size_inputs}
\alias{create_effect_size_inputs}
\title{Create Effect Size Input Components}
\usage{
create_effect_size_inputs()
}
\description{
Create Effect Size Input Components
}
\keyword{internal}
