% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test_registry.R
\name{create_prop_2groups_spec}
\alias{create_prop_2groups_spec}
\title{Two Proportions Specification}
\usage{
create_prop_2groups_spec()
}
\description{
Two Proportions Specification
}
\keyword{internal}
