% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_components.R
\name{create_sample_size_inputs}
\alias{create_sample_size_inputs}
\title{Create Sample Size Input Components}
\usage{
create_sample_size_inputs()
}
\description{
Create Sample Size Input Components
}
\keyword{internal}
