% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test_registry.R
\name{create_ttest_2groups_spec}
\alias{create_ttest_2groups_spec}
\title{Two-Group Independent t-test Specification}
\usage{
create_ttest_2groups_spec()
}
\description{
Two-Group Independent t-test Specification
}
\keyword{internal}
