% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test_registry.R
\name{create_ttest_one_sample_spec}
\alias{create_ttest_one_sample_spec}
\title{One-Sample t-test Specification}
\usage{
create_ttest_one_sample_spec()
}
\description{
One-Sample t-test Specification
}
\keyword{internal}
