% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test_registry.R
\name{create_ttest_paired_spec}
\alias{create_ttest_paired_spec}
\title{Paired t-test Specification}
\usage{
create_ttest_paired_spec()
}
\description{
Paired t-test Specification
}
\keyword{internal}
