% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_components.R
\name{create_ui}
\alias{create_ui}
\title{Create UI Components for the Shiny Application}
\usage{
create_ui()
}
\description{
Internal functions to create the user interface components.
}
\keyword{internal}
