% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_server_factory.R
\name{.format_generic_html_report}
\alias{.format_generic_html_report}
\title{Format Generic HTML Report}
\usage{
.format_generic_html_report(report_data, test_spec)
}
\description{
Format Generic HTML Report
}
\keyword{internal}
