% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_server_factory.R
\name{.format_generic_text_report}
\alias{.format_generic_text_report}
\title{Format Generic Text Report}
\usage{
.format_generic_text_report(report_data, test_spec)
}
\description{
Format Generic Text Report
}
\keyword{internal}
