% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test_registry.R
\name{get_power_test_registry}
\alias{get_power_test_registry}
\title{Get the complete test registry}
\usage{
get_power_test_registry()
}
\value{
List of all available power tests with their configurations
}
\description{
Get the complete test registry
}
\keyword{internal}
