% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_helpers.R
\name{get_prop_axis_label}
\alias{get_prop_axis_label}
\title{Get X-axis Label for Proportion Power Curves}
\usage{
get_prop_axis_label(method)
}
\arguments{
\item{method}{Effect size method}
}
\value{
Axis label
}
\description{
Returns a formatted axis label for the power curve plot
}
\keyword{internal}
