% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_helpers.R
\name{prop_to_cohens_h}
\alias{prop_to_cohens_h}
\title{Calculate Cohen's h from Two Proportions}
\usage{
prop_to_cohens_h(p1, p2)
}
\arguments{
\item{p1}{Proportion in group 1 (0 to 1)}

\item{p2}{Proportion in group 2 (0 to 1)}
}
\value{
Cohen's h effect size
}
\description{
Cohen's h is the standardized difference between two proportions:
h = 2 * (arcsin(sqrt(p1)) - arcsin(sqrt(p2)))
}
\examples{
\dontrun{
prop_to_cohens_h(0.6, 0.4)  # Returns ~0.41
}
}
\keyword{internal}
