% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ZZPOWER_CONSTANTS}
\alias{ZZPOWER_CONSTANTS}
\title{zzpower Package Constants}
\format{
An object of class \code{list} of length 66.
}
\usage{
ZZPOWER_CONSTANTS
}
\description{
Centralized constants for the zzpower package. These values define
parameter bounds, defaults, and computational settings.
}
\keyword{internal}
